requirements.txt
python-telegram-bot==13.6
pandas
pytz
openpyxl
dotenv
